requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.1
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.2
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   google-genai
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via aiohttp
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via uvicorn
dataclasses-json==0.6.7
    # via langchain-community
et-xmlfile==2.0.0
    # via openpyxl
fastapi==0.120.0
    # via -r requirements.in
filetype==1.2.0
    # via langchain-google-genai
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
google-ai-generativelanguage==0.9.0
    # via langchain-google-genai
google-api-core[grpc]==2.27.0
    # via google-ai-generativelanguage
google-auth==2.41.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-genai
google-genai==1.46.0
    # via -r requirements.in
googleapis-common-protos==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   google-genai
    #   langgraph-sdk
    #   langsmith
    #   ollama
httpx-sse==0.4.3
    # via langchain-community
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
joblib==1.5.2
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
kaleido==0.2.1
    # via -r requirements.in
kneed==0.8.5
    # via -r requirements.in
langchain==1.0.2
    # via -r requirements.in
langchain-classic==1.0.0
    # via langchain-community
langchain-community==0.4
    # via -r requirements.in
langchain-core==1.0.0
    # via
    #   langchain
    #   langchain-classic
    #   langchain-community
    #   langchain-google-genai
    #   langchain-ollama
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-genai==3.0.0
    # via -r requirements.in
langchain-ollama==1.0.0
    # via -r requirements.in
langchain-text-splitters==1.0.0
    # via
    #   -r requirements.in
    #   langchain-classic
langgraph==1.0.1
    # via langchain
langgraph-checkpoint==3.0.0
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.1
    # via langgraph
langgraph-sdk==0.2.9
    # via langgraph
langsmith==0.4.38
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
markdown==3.9
    # via -r requirements.in
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
narwhals==2.9.0
    # via plotly
numpy==2.3.4
    # via
    #   -r requirements.in
    #   kneed
    #   langchain-community
    #   pandas
    #   scikit-learn
    #   scipy
    #   xgboost
ollama==0.6.0
    # via
    #   -r requirements.in
    #   langchain-ollama
openpyxl==3.1.5
    # via -r requirements.in
orjson==3.11.3
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.11.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   plotly
pandas==2.3.3
    # via -r requirements.in
pillow==12.0.0
    # via -r requirements.in
plotly==6.3.1
    # via -r requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==6.33.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyarrow==21.0.0
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.3
    # via
    #   fastapi
    #   google-genai
    #   langchain
    #   langchain-classic
    #   langchain-core
    #   langchain-google-genai
    #   langgraph
    #   langsmith
    #   ollama
    #   pydantic-settings
pydantic-core==2.41.4
    # via pydantic
pydantic-settings==2.11.0
    # via langchain-community
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   uvicorn
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   langchain-classic
    #   langchain-community
    #   langchain-core
    #   uvicorn
requests==2.32.5
    # via
    #   google-api-core
    #   google-genai
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
scikit-learn==1.7.2
    # via -r requirements.in
scipy==1.16.2
    # via
    #   kneed
    #   scikit-learn
    #   xgboost
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via
    #   langchain-classic
    #   langchain-community
starlette==0.48.0
    # via fastapi
tenacity==9.1.2
    # via
    #   google-genai
    #   langchain-community
    #   langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   fastapi
    #   google-genai
    #   grpcio
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.38.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
xgboost==3.1.1
    # via -r requirements.in
xxhash==3.6.0
    # via langgraph
yarl==1.22.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
